Convert settings plugin to Java
Use Java (rather than Kotlin) for the settings plugin so that it can exist in the same project as the project conventions.
It might not have been necessary, but`kotlin-dsl` is required for the project conventions - but this was troublesome when applied to the settings plugin in Kotlin, and I couldn't figure out how to get it working.